trying to detect broken nodes
We will do this as when the end user clicks a link to a local server
then we will make a note that we tried to contact an intranet server by
using a tracking code, which is generated by web browser. If the
connection succeeds then the main server will send the tracking code
back to the client with a Success method. This means that we can detect
which servers are running but not contactable, usually because of
firewall restrictions.