Show a basic tree for history mode
I have started to think about how the history mode will look for a file.
I think it makes sense to show it as a tree, with older nodes to the
left and more recent nodes to the right. This could change of course,
but I have to start somewhere