Fixed Windows issue
when refactoring the code VisiFile would fail to start on Windows. The
issue was that native modules must be copied to the local filesystem
BEFORE being referred to by NodeJS in a "require" statement

I fixed this last night and I just tested that the .exe works as I use
the .exe at work. So I checked the code in just now :)