Fixed a bug in the electron app
The homepage would show through the editor. This only happened when running in Electron. I thought it was a Chrome bug at first, so I tried unsuccessfully to get a more up to date version of Chrome, but then found it too much work since I had to upgrade Electron to do this, so I put a quick hack in which uses z-index to move the homepage to the background