Added a hack to delete the Ace editor
When switching from the Ace editor to the Help the Ace editor is not always destroyed. This is related to how the Ace editor parent DOM node is destroyed by VueJS. This was a bad bug as it happened when I was demoing Yazz to someone at Microsoft!!! They were NOT impressed!