Trying to make the Docker control a container
We thought about how to give a default UI to a non GUI control a lot. At first we cam up with a lot of really complicated schemes to have embedded control tags within objects, but then we found out that we could try to use out existing grouping functionality to allow for example a UI widget to be placed in a non UI container.

Time will tell if this works well, but it "seems" elegant

PT