REST Component: Making a second version of the Rest API call
We are makign a new version of the REST API control so that we can have better API testing from within the REST component. Currently we have to have everything live, but this should allow us to test stuff, and then make it live when the API that we are testing works. So we are adding a new function for this, callRestApiInternalV2! Boy, do I love our version naming scheme, NOT!