Added code_tag_v2
"code_tag_v2" will supercede "code_tag". "code_tag" used to have a concept of the "LATEST" code tagged, but this doesn't work in a multiplayer world with many code branches. The LATEST code will be determined in a different table going forward, and "code_tag_v2" will be used to identify the multiple "TIP"s of the code