Can write "Yazz.component" instead of "Vue.component"
Only a small change, but this is the first step in the process of using the commit ID instead of the Base Component ID to identify UI components. The benefit that this will give us is that it will allow us to have different versions of the same UI component in an app at the same time.