Remove unused code for allowing end user apps to be edited
I can't even remember the last time that this actually worked. Maybe we can rebuild it in the future with an "Edit" button that then downloads the code somehow from a file system or IPFS, but for now it just complicates the code. The next few checkouts will be cleanup code too, some of it related to this